🚚rename: [자소서] 폴더 구조 변경
- (personal-statement) 하위에 공통 _components 폴더를 두고, 하위에 페이지별 폴더를 만드는 구조로 변경

KAN-106
