+++ src/app/(route)/(personal-statement)/_components/_editing/PSEditingBox.tsx
