+++ src/app/(route)/(personal-statement)/_components/_read/PSReadContainer.tsx
@@ -3,13 +3,13 @@
 import { useEffect, useState } from 'react';
 import { useRouter } from 'next/navigation';
 import { usePDF } from 'react-to-pdf';
-import PSFooter from '../../_components/PSFooter';
-import PSReadContent from './PSReadContent';
-import { PSFormData } from '../../create/_types/psCreate';
 import psReadTest from './psReadTest.json';
-import PSHeader from '../../_components/PSHeader';
-import BackButton from '../../../../../../_components/BackButton';
-import Modal, { initailModalState } from '../../../../../../_components/Modal';
+import BackButton from '../../../../_components/BackButton';
+import Modal, { initailModalState } from '../../../../_components/Modal';
+import { PSFormData } from '../../_types/psCreate';
+import PSFooter from '../PSFooter';
+import PSHeader from '../PSHeader';
+import PSReadContent from './PSReadContent';
 
 export default function PSReadContainer() {
   const router = useRouter();
